# IGNITE-24196 Sql. Merge projection into join node