https://issues.apache.org/jira/browse/IGNITE-24196

This patch introduces new `SqlJoinProjection` expression to use instead of `RowFactory#concat` in join nodes. Besides, in reorganizes schema initialization in `ItJoinOrderTest` to take an advantage of DDL batching.

Results of SqlJoinBenchmark on my laptop as follow:
```
Before
Benchmark                        (clusterSize)  (fsync)  Mode  Cnt    Score   Error  Units
SqlJoinBenchmark.leftHashJoin                1    false  avgt   20  122.215 ± 0.768  ms/op
SqlJoinBenchmark.leftMergeJoin               1    false  avgt   20  167.923 ± 1.183  ms/op
SqlJoinBenchmark.leftNestedJoin              1    false  avgt   20  208.322 ± 1.220  ms/op

After Projection Fusion
Benchmark                        (clusterSize)  (fsync)  Mode  Cnt    Score   Error  Units
SqlJoinBenchmark.leftHashJoin                1    false  avgt   20   72.667 ± 0.647  ms/op
SqlJoinBenchmark.leftMergeJoin               1    false  avgt   20  138.896 ± 1.919  ms/op
SqlJoinBenchmark.leftNestedJoin              1    false  avgt   20  173.719 ± 0.852  ms/op

After Projection Fusion + Field getter optimization
Benchmark                        (clusterSize)  (fsync)  Mode  Cnt    Score   Error  Units
SqlJoinBenchmark.leftHashJoin                1    false  avgt   20   54.473 ± 0.099  ms/op
SqlJoinBenchmark.leftMergeJoin               1    false  avgt   20  115.621 ± 0.931  ms/op
SqlJoinBenchmark.leftNestedJoin              1    false  avgt   20  111.341 ± 0.250  ms/op
```

Also I run query 64 from TPC-DS with sf=1 to make sure there is no degradation:
```
Before
Benchmark           (fsync)  (queryId)  Mode  Cnt     Score     Error  Units
TpcdsBenchmark.run    false         64  avgt   20  5046.916 ± 150.744  ms/op

After  Projection Fusion + Field getter optimization
Benchmark           (fsync)  (queryId)  Mode  Cnt     Score     Error  Units
TpcdsBenchmark.run    false         64  avgt   20  4799.201 ± 160.410  ms/op
```
------------
Thank you for submitting the pull request.

To streamline the review process of the patch and ensure better code quality
we ask both an author and a reviewer to verify the following:

### The Review Checklist
- [ ] **Formal criteria:** TC status, codestyle, mandatory documentation. Also make sure to complete the following:  
\- There is a single JIRA ticket related to the pull request.  
\- The web-link to the pull request is attached to the JIRA ticket.  
\- The JIRA ticket has the Patch Available state.  
\- The description of the JIRA ticket explains WHAT was made, WHY and HOW.  
\- The pull request title is treated as the final commit message. The following pattern must be used: IGNITE-XXXX Change summary where XXXX - number of JIRA issue.
- [ ] **Design:** new code conforms with the design principles of the components it is added to.
- [ ] **Patch quality:** patch cannot be split into smaller pieces, its size must be reasonable.
- [ ] **Code quality:** code is clean and readable, necessary developer documentation is added if needed.
- [ ] **Tests code quality:** test set covers positive/negative scenarios, happy/edge cases. Tests are effective in terms of execution time and resources.

### Notes
- [Apache Ignite Coding Guidelines](https://cwiki.apache.org/confluence/display/IGNITE/Java+Code+Style+Guide)